Update checkzero.c
typedef bool int; is missing, Try to use return at the end of function or outside the loops and not inside the loops.
Added typedef to create new datatype named bool which is same as int. Used break statement to stop the loop which reduces the time complexity if number contains zero in it.